+++ src/Service/Generator/ControllerClassGenerator.php
@@ -6,16 +6,12 @@ use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Core\Entity\EntityFieldManagerInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\File\FileSystemInterface;
-use Drupal\wmmodel\Factory\ModelFactory;
 use Drupal\wmscaffold\Service\Helper\IdentifierNaming;
 use Drupal\wmscaffold\Service\Helper\StringCapitalisation;
 use PhpParser\Node\Stmt;
 
 class ControllerClassGenerator extends ClassGeneratorBase
 {
-    /** @var ModelFactory */
-    protected $modelFactory;
-
     /** @var string */
     protected $baseClass;
 
@@ -23,11 +19,9 @@ class ControllerClassGenerator extends ClassGeneratorBase
         EntityTypeManagerInterface $entityTypeManager,
         EntityFieldManagerInterface $entityFieldManager,
         FileSystemInterface $fileSystem,
-        ConfigFactoryInterface $configFactory,
-        ModelFactory $modelFactory
+        ConfigFactoryInterface $configFactory
     ) {
         parent::__construct($entityTypeManager, $entityFieldManager, $fileSystem, $configFactory);
-        $this->modelFactory = $modelFactory;
 
         $this->baseClass = $this->config->get('generators.controller.base_class');
     }
@@ -36,9 +30,8 @@ class ControllerClassGenerator extends ClassGeneratorBase
     {
         $className = $this->buildClassName($entityType, $bundle, $module, true);
         $namespaceName = $this->buildNamespaceName($entityType, $module);
-        $definition = $this->entityTypeManager->getDefinition($entityType);
         $modelClass = new \ReflectionClass(
-            $this->modelFactory->getClassName($definition, $bundle)
+            $this->entityTypeManager->getStorage($entityType)->getEntityClass($bundle)
         );
 
         $variableName = StringCapitalisation::toCamelCase($modelClass->getShortName());